+++ include/game/sdl/text.h
@@ -1,6 +1,6 @@
-#include <SDL_pixels.h>
-#include <SDL_render.h>
-#include <SDL_ttf.h>
+#include <SDL3/SDL_pixels.h>
+#include <SDL3/SDL_render.h>
+#include <SDL3_ttf/SDL_ttf.h>
 #include <cstdint>
 #include <string>
 
@@ -21,9 +21,9 @@ namespace game::sdl {
     void draw_text(SDL_Renderer* renderer,
                    TTF_Font* font,
                    const std::string& text,
-                   int x,
-                   int y,
-                   int font_size,
+                   float x,
+                   float y,
+                   float font_size,
                    HAlign h_align,
                    VAlign v_align,
                    SDL_Color color);
